Move transcription logic to speech inference provider
Added a DynamicRangeLoading class
This class handles loading of ranges.
If the user starts a video it starts to load 0-9 sections.
However, the user switches to section 5 so the loading should continue
from 5-9 and then finish what is left (0-4)

This is rather complicated logic, but this will be useful for video
inferencing too.