Add FixedGrid widget for fixed size tiles
The issue was that the sliver grid was resizing the tiles
This made the UI so much harder to design.
Now instead of that we pad to the right in the grid and the tiles are
fixed width.
When the layout changes the amount of tiles shown side by side is
calculated.