Add translation toggle to whisper  (#60)
* Add language selector

Won't keep this since whisper only allows translation to english.
Transcribing works for those other languages. Just not any language to
other than english.

* Remove language selector and add toggle for translation to english

Whisper models are not trained to translate to any language.
They either transcribe in a language, or translate to english.

* Set language in speech to text to always set language

This fixes issue when language is set it is kept.

* Test commit for language setting in speech to text

* Fix whisper Use AVSEEK_FLAG_BACKWARD to seek more accurately

We were getting different outputs. This was caused by the audio grabber.
It still isnt perfectly grabbing the same data, but better now

* Keep current location when toggling translation