Change routing to use push and pop
Previously this was .go(). This was done to fix the providers to
properly dispose.
This was not neccesary. The main navigation buttons need to .go(), but
others can keep stack just fine.

Also added a dialog for when closing during download to make it clear to
the user that download is stopped.