Fixes state issues with camera and providers
Show model loading animation during transition

GraphRunner::closeCamera can be triggered twice which might cause a race
condition. Instead of fixing this properly with a mutex in C++ I just
eat the error code.

If there was a previous provider make sure we close the camera before loading a new one.
This ensures that the camera is ready to be used again when the model is loaded.